requirements.txt
python-dotenv>=0.19.0
openai>=1.3.0
pandas>=1.3.0
numpy>=1.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
umap-learn>=0.5.0
scikit-learn>=0.24.0
plotly>=5.3.0
rich>=10.0.0
tqdm>=4.62.0
matplotlib>=3.4.0
